modules/image: remove unused parameters and fix signedness
Unused parameters in static functions are removed. Unused parameters in
public functions or X11 callbacks can't be removed because they impact
the command-line or are required even if unused, so a "foo=foo"
workaround is inserted to fool the compiler.

Stick to the structure definitions about structure members signedness
(some of them should be changes, images width/heith is unsigned in fact,
but changing it would break the API).

+++ modules/src/image/level_lines/cll_remove.c
@@ -67,7 +67,11 @@ Cmimage cll_remove(Cmimage cmimage, int *L)
 	mwerror(FATAL,1,"Input cmimage contains mixed types of level lines !\n");
       ll_next = ll->next;
       /*  check_ll(ll);*/
-      if (mw_length_cmorpho_line(ll) < *L) {remove_cll(cmimage,ll); Nrm++; }
+      if (mw_length_cmorpho_line(ll) < (unsigned int) *L)
+      {
+	remove_cll(cmimage,ll);
+	Nrm++;
+      }
     }
 
   if (ll_type == 0)
